+++ huginn-net-db/src/database/mod.rs
@@ -2,25 +2,25 @@
 
 mod collection;
 mod constants;
-mod label;
-#[cfg(any(feature = "tcp", feature = "http"))]
-mod keys;
-#[cfg(feature = "http")]
-mod http;
-#[cfg(feature = "tcp")]
-mod tcp;
 #[cfg(all(feature = "tcp", feature = "http"))]
 mod full;
+#[cfg(feature = "http")]
+mod http;
+#[cfg(any(feature = "tcp", feature = "http"))]
+mod keys;
+mod label;
+#[cfg(feature = "tcp")]
+mod tcp;
 
 pub use collection::FingerprintCollection;
-pub use label::{Label, Type};
-#[cfg(feature = "tcp")]
-pub use keys::TcpIndexKey;
+#[cfg(all(feature = "tcp", feature = "http"))]
+pub use full::Database;
+#[cfg(feature = "http")]
+pub use http::HttpDatabase;
 #[cfg(feature = "http")]
 pub use keys::HttpIndexKey;
 #[cfg(feature = "tcp")]
+pub use keys::TcpIndexKey;
+pub use label::{Label, Type};
+#[cfg(feature = "tcp")]
 pub use tcp::TcpDatabase;
-#[cfg(feature = "http")]
-pub use http::HttpDatabase;
-#[cfg(all(feature = "tcp", feature = "http"))]
-pub use full::Database;
