+++ huginn-net-db/src/http/mod.rs
@@ -6,6 +6,8 @@ pub use huginn_net_http::http::{
     HttpDiagnosis, Version,
 };
 
+mod distances;
 mod signature;
 
+pub use distances::{distance_expsw, distance_header, distance_http_version};
 pub use signature::{HttpMatchQuality, Signature};
