+++ huginn-net-db/src/matcher/http_signature_matcher.rs
@@ -1,7 +1,7 @@
 use crate::database::{HttpDatabase, Label, Type};
 use crate::db_matching_trait::FingerprintDb;
-use crate::observable_signals::{HttpRequestObservation, HttpResponseObservation};
 use huginn_net_http::matcher_api::{HttpMatcher, HttpRequestMatch, HttpResponseMatch, UaOsMatch};
+use huginn_net_http::observable::{HttpRequestObservation, HttpResponseObservation};
 use huginn_net_http::output::{Browser, OsKind, WebServer};
 use std::sync::Arc;
