+++ huginn-net-db/src/tcp/signature.rs
@@ -1,4 +1,6 @@
 use crate::db_matching_trait::MatchQuality;
+use core::fmt;
+use std::fmt::Formatter;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum TcpMatchQuality {
@@ -56,4 +58,91 @@ pub struct Signature {
     pub quirks: Vec<super::Quirk>,
     /// payload size classification
     pub pclass: super::PayloadSize,
+}
+
+// Shared p0f format: `version:ittl:olen:mss:wsize,wscale:olayout:quirks:pclass`.
+trait TcpDisplayFormat {
+    fn get_version(&self) -> super::IpVersion;
+    fn get_ittl(&self) -> super::Ttl;
+    fn get_olen(&self) -> u8;
+    fn get_mss(&self) -> Option<u16>;
+    fn get_wsize(&self) -> super::WindowSize;
+    fn get_wscale(&self) -> Option<u8>;
+    fn get_olayout(&self) -> &[super::TcpOption];
+    fn get_quirks(&self) -> &[super::Quirk];
+    fn get_pclass(&self) -> super::PayloadSize;
+
+    fn format_tcp_display(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        write!(f, "{}:{}:{}:", self.get_version(), self.get_ittl(), self.get_olen())?;
+
+        if let Some(mss) = self.get_mss() {
+            write!(f, "{mss}")?;
+        } else {
+            f.write_str("*")?;
+        }
+
+        write!(f, ":{},", self.get_wsize())?;
+
+        if let Some(scale) = self.get_wscale() {
+            write!(f, "{scale}")?;
+        } else {
+            f.write_str("*")?;
+        }
+
+        f.write_str(":")?;
+
+        for (i, o) in self.get_olayout().iter().enumerate() {
+            if i > 0 {
+                f.write_str(",")?;
+            }
+            write!(f, "{o}")?;
+        }
+
+        f.write_str(":")?;
+
+        for (i, q) in self.get_quirks().iter().enumerate() {
+            if i > 0 {
+                f.write_str(",")?;
+            }
+            write!(f, "{q}")?;
+        }
+
+        write!(f, ":{}", self.get_pclass())
+    }
+}
+
+impl TcpDisplayFormat for Signature {
+    fn get_version(&self) -> super::IpVersion {
+        self.version
+    }
+    fn get_ittl(&self) -> super::Ttl {
+        self.ittl.clone()
+    }
+    fn get_olen(&self) -> u8 {
+        self.olen
+    }
+    fn get_mss(&self) -> Option<u16> {
+        self.mss
+    }
+    fn get_wsize(&self) -> super::WindowSize {
+        self.wsize.clone()
+    }
+    fn get_wscale(&self) -> Option<u8> {
+        self.wscale
+    }
+    fn get_olayout(&self) -> &[super::TcpOption] {
+        &self.olayout
+    }
+    fn get_quirks(&self) -> &[super::Quirk] {
+        &self.quirks
+    }
+    fn get_pclass(&self) -> super::PayloadSize {
+        self.pclass
+    }
+}
+
+impl fmt::Display for Signature {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        self.format_tcp_display(f)
+    }
 }
